fix(templates): ComponentDetailView is Design Documentation
- Design Documentation = component detail page (Button with live
  preview, usage, best practices, code examples)
- Technical Documentation = Getting Started + component category grid

Made-with: Cursor
